Move files to add layers and structure
