+++ lib/service/push_notifications/push_notifications.dart
@@ -4,7 +4,7 @@ import 'dart:io';
 
 import 'package:corona_trace/main.dart';
 import 'package:corona_trace/network/api_repository.dart';
-import 'package:corona_trace/network/repository_notifications.dart';
+import 'package:corona_trace/network/notification/response_notification_item.dart';
 import 'package:corona_trace/ui/notifications/ct_notification_map_detail.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
