+++ lib/ui/notifications/ct_notification_detail_card.dart
@@ -1,5 +1,5 @@
 import 'package:corona_trace/app_constants.dart';
-import 'package:corona_trace/network/repository_notifications.dart';
+import 'package:corona_trace/network/notification/response_notification_item.dart';
 import 'package:corona_trace/ui/notifications/ct_notification_item.dart';
 import 'package:corona_trace/utils/app_localization.dart';
 import 'package:flutter/material.dart';
