+++ lib/ui/screens/getting_started_screen.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:corona_trace/location_updates.dart';
 import 'package:corona_trace/network/api_repository.dart';
-import 'package:corona_trace/push_notifications/push_notifications.dart';
+import 'package:corona_trace/service/push_notifications/push_notifications.dart';
 import 'package:corona_trace/ui/base_state.dart';
 import 'package:corona_trace/ui/screens/onboarding.dart';
 import 'package:corona_trace/ui/screens/user_info_collector_screen.dart';
